refactor: 코드 리팩토링
Position 미사용 생성자 삭제, 생성자 package-access로 수정
Positions Position의 생성을 막기 위해 패키지 분리
ALL 패키지 변경에 따른 리팩토링
