+++ src/main/java/chess/domain/piece/movable/BlockedMovable.java
@@ -2,7 +2,7 @@ package chess.domain.piece.movable;
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
-import chess.domain.position.Positions;
+import chess.domain.position.positions.Positions;
 import java.util.List;
 public class BlockedMovable implements Movable {
     private final Directions moveDirections;
