+++ src/main/java/chess/domain/piece/movable/UnblockedMovable.java
@@ -2,7 +2,7 @@ package chess.domain.piece.movable;
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
-import chess.domain.position.Positions;
+import chess.domain.position.positions.Positions;
 import java.util.List;
 import java.util.stream.Collectors;
 public class UnblockedMovable implements Movable {
