+++ src/main/java/chess/domain/util/WrongOperationException.java
@@ -1,11 +1,11 @@
 package chess.domain.util;
 
 public class WrongOperationException extends RuntimeException {
-    public WrongOperationException() {
-        this("수행할 수 없는 명령입니다.");
-    }
-
     public WrongOperationException(String message) {
         super(message);
+    }
+
+    public WrongOperationException() {
+        this("수행할 수 없는 명령입니다.");
     }
 }
